airflow-docker/requirements.txt
kaggle
pandas
google-cloud-storage
python-dotenv
pyspark
apache-airflow-providers-google
